+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs/DamageDoneToPartyMembers.lua
@@ -7,6 +7,17 @@ UI
 @section UI
 ]]
 
+local L = LibStub("AceLocale-3.0"):GetLocale("MyDungeonsBook");
+
+local function getDamageToEachPartyMemberSpellMenu(rows, index, cols)
+	local spellId = rows[index].cols[1].value;
+	local report = MyDungeonsBook:DamageDoneToPartyMembersFrame_Report_Create(rows[index], cols);
+	return {
+		MyDungeonsBook:WowHead_Menu_Spell(spellId),
+		MyDungeonsBook:Report_Menu(report)
+	};
+end
+
 --[[--
 Create a frame for Damage Done To Party Members tab (data is taken from `mechanics[ALL-DAMAGE-DONE-TO-PARTY-MEMBERS]`).
 
@@ -23,6 +34,13 @@ function MyDungeonsBook:DamageDoneToPartyMembersFrame_Create(parentFrame, challe
 	local table = self:TableWidget_Create(columns, 11, 40, nil, damageDoneToPartyMembersFrame, "damage-done-to-party-members");
 	table:SetData(data);
 	table:RegisterEvents({
+		OnClick = function(_, _, data, _, _, realrow, _, _, button)
+			if (button == "RightButton" and realrow) then
+				if (data[realrow].cols[1].value > 0) then
+					EasyMenu(getDamageToEachPartyMemberSpellMenu(data, realrow, table.cols), self.menuFrame, "cursor", 0 , 0, "MENU");
+				end
+			end
+		end,
 		OnEnter = function (_, cellFrame, data, _, _, realrow, column, _)
 			if (realrow) then
 				if (column == 2 or column == 3) then
@@ -215,4 +233,15 @@ function MyDungeonsBook:DamageDoneToPartyMembersFrame_GetAvoidableSummaryRow(cha
 		end
 	end
 	return tableData;
+end
+
+--[[--
+@param[type=table] row
+@param[type=table] cols
+@return[type=table]
+]]
+function MyDungeonsBook:DamageDoneToPartyMembersFrame_Report_Create(row, cols)
+	local spellLink = GetSpellLink(row.cols[1].value);
+	local title = string.format(L["MyDungeonsBook Damage taken by party from %s:"], spellLink);
+	return self:Table_PlayersRow_Report_Create(row, cols, {4, 6, 8, 10, 12, 14}, title);
 end