+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/EffectsAndAuras/Tabs/AllBuffsAndDebuffsOnPartyMembers/Tabs/AllBuffsAndDebuffsOnPartyMember.lua
@@ -8,7 +8,13 @@ UI
 ]]
 
 local L = LibStub("AceLocale-3.0"):GetLocale("MyDungeonsBook");
-local AceGUI = LibStub("AceGUI-3.0");
+
+local function getAllBuffsAndDebuffsSpellMenu(spellId)
+    return {
+        MyDungeonsBook:WowHead_Menu_Spell(spellId)
+    };
+end
+
 
 --[[--
 @param[type=Frame] parentFrame
@@ -23,6 +29,11 @@ function MyDungeonsBook:AllBuffsAndDebuffsOnPartyMemberFrame_Create(parentFrame,
     local buffsTable = self:TableWidget_Create(columns, 10, 40, nil, allBuffsAndDebuffsOnPartyMemberFrame, "party-members-auras-on-" .. unitId .. "-buffs");
     buffsTable:SetData(buffsData);
     buffsTable:RegisterEvents({
+        OnClick = function(_, _, data, _, _, realrow, _, _, button)
+            if (button == "RightButton" and realrow) then
+                EasyMenu(getAllBuffsAndDebuffsSpellMenu(data[realrow].cols[1].value), self.menuFrame, "cursor", 0 , 0, "MENU");
+            end
+        end,
         OnEnter = function (_, cellFrame, data, _, _, realrow, column)
             if (realrow) then
                 if (column == 2 or column == 3) then
@@ -42,6 +53,11 @@ function MyDungeonsBook:AllBuffsAndDebuffsOnPartyMemberFrame_Create(parentFrame,
     local debuffsData = self:AllBuffsAndDebuffsOnPartyMemberFrame_GetDataForTable(challengeId, "PARTY-MEMBERS-AURAS", unitId, "DEBUFF");
     debuffsTable:SetData(debuffsData);
     debuffsTable:RegisterEvents({
+        OnClick = function(_, _, data, _, _, realrow, _, _, button)
+            if (button == "RightButton" and realrow) then
+                EasyMenu(getAllBuffsAndDebuffsSpellMenu(data[realrow].cols[1].value), self.menuFrame, "cursor", 0 , 0, "MENU");
+            end
+        end,
         OnEnter = function (_, cellFrame, data, _, _, realrow, column)
             if (realrow) then
                 if (column == 2 or column == 3) then