+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/EffectsAndAuras/Tabs/AvoidableDebuffs.lua
@@ -7,6 +7,17 @@ UI
 @section UI
 ]]
 
+local L = LibStub("AceLocale-3.0"):GetLocale("MyDungeonsBook");
+
+local function getAvoidableDebuffsSpellMenu(rows, index, cols)
+	local spellId = rows[index].cols[1].value;
+	local report = MyDungeonsBook:AvoidableDebuffsFrame_Report_Create(rows[index], cols);
+	return {
+		MyDungeonsBook:WowHead_Menu_Spell(spellId),
+		MyDungeonsBook:Report_Menu(report)
+	};
+end
+
 --[[--
 Create a frame for Avoidable Debuffs tab (data is taken from `mechanics[**-AVOIDABLE-AURAS]`).
 
@@ -23,6 +34,11 @@ function MyDungeonsBook:AvoidableDebuffsFrame_Create(parentFrame, challengeId)
 	local table = self:TableWidget_Create(columns, 11, 40, nil, avoidableDebuffsFrame, "avoidable-debuffs");
 	table:SetData(data);
 	table:RegisterEvents({
+		OnClick = function(_, _, data, _, _, realrow, _, _, button)
+			if (button == "RightButton" and realrow) then
+				EasyMenu(getAvoidableDebuffsSpellMenu(data, realrow, table.cols), self.menuFrame, "cursor", 0 , 0, "MENU");
+			end
+		end,
 		OnEnter = function (_, cellFrame, data, _, _, realrow, column)
 			if (realrow) then
 				if (column == 2 or column == 3) then
@@ -101,4 +117,15 @@ function MyDungeonsBook:AvoidableDebuffsFrame_GetDataForTable(challengeId, key)
 		tinsert(remappedTableData, r);
 	end
 	return remappedTableData;
+end
+
+--[[--
+@param[type=table] row
+@param[type=table] cols
+@return[type=table]
+]]
+function MyDungeonsBook:AvoidableDebuffsFrame_Report_Create(row, cols)
+	local spellLink = GetSpellLink(row.cols[1].value);
+	local title = string.format(L["MyDungeonsBook Debuff %s hits:"], spellLink);
+	return self:Table_PlayersRow_Report_Create(row, cols, {4, 5, 6, 7, 8, 9}, title);
 end